// src/interfaces/person.interface.ts
export interface Person {
    nombre: string;
    apellido_paterno: string;
    apellido_materno: string;
    fecha_nacimiento: Date;
    telefono: string;
    direccion: string;
    numero_documento: string;
    id_tip_documento: number;
    id_genero: number;
    id_estado_civil: number;
    id_pais: number;
    id_region: number;
    id_provincia: number;
    id_comuna: number;
    id_estado_usuario: number;
    id_rol_usuario: number;
}